subscription (toggled off), new settings design

+++ src/components/subscription/PaymentMethods.vue
@@ -1,19 +1,21 @@
 <template>
 	<div class="flex flex-col h-full">
 		<div class="flex select-none">
-			<div class="py-2 px-4 bg-info cursor-default">Payment Methods</div>
+			<div class="py-2 px-4 bg-info cursor-default" data-testid="payment-methods-view">Payment Methods</div>
 		</div>
-		<div class="bg-info w-full h-2 mb-2"></div>
-		<div class="flex flex-col gap-3 overflow-y-auto" data-testid="payment-methods-view">
+		<div class="bg-info h-2 mb-2 mr-2"></div>
+		<div v-if="!methods?.length" class="m-5" data-testid="payment-methods-none">No Payments Methods Yet</div>
+		<div class="flex flex-col gap-3 overflow-y-auto">
 			<template v-for="method of methods" :key="method.id">
 				<PaymentMethodItem
 					:method="method"
 					:is-default="method.id === defaultMethod?.id"
+					:show-actions="true"
 					@make-default="makeDefault(method)"
 					@delete="deleteMethod(method)"
 				></PaymentMethodItem>
 			</template>
-			<div class="flex justify-end w-80 pt-10 pb-10">
+			<div v-if="showCreate" class="flex justify-end w-80 pt-10 pb-10">
 				<button @click="createNew">Create New</button>
 			</div>
 		</div>
@@ -24,12 +26,23 @@
 import { onMounted, ref } from 'vue'
 import type { PaymentMethod } from '../../services/types/subscription'
 import PaymentMethodItem from './PaymentMethodItem.vue'
-import { noteManager } from '../../global'
+import { deletePaymentMethodDialog, noteManager } from '../../global'
+import { emptyGuid } from '../../services/types/guid'
+
+const emit = defineEmits(['pay-in-progress'])
 
 const methods = ref<PaymentMethod[]>()
 const defaultMethod = ref<PaymentMethod>()
+const showCreate = ref(false)
 
 const populate = async () => {
+	try {
+		const customer = await noteManager.paymentClient.getCustomerData()
+		showCreate.value = !!customer
+	} catch {
+		showCreate.value = false
+	}
+
 	const items = await noteManager.paymentClient.getPaymentMethods()
 	methods.value = items
 	let def = items[0]
@@ -48,12 +61,17 @@ const makeDefault = async (method: PaymentMethod) => {
 	await populate()
 }
 const deleteMethod = async (method: PaymentMethod) => {
-	await noteManager.paymentClient.deletePaymentMethodDefault(method.id)
-	await populate()
+	deletePaymentMethodDialog.value.show(`${method.brand}, ${method.name}`, async confirmed => {
+		if (confirmed) {
+			await noteManager.paymentClient.deletePaymentMethodDefault(method.id)
+			await populate()
+		}
+	})
 }
 
 const createNew = async () => {
 	const result = await noteManager.paymentClient.createNewPaymentMethod({ clientReference: 'create-method' })
+	emit('pay-in-progress', emptyGuid(), true, result.link, methods.value.length + 1)
 	window.open(result.link, '_blank')
 }
 </script>
